requirements.txt
streamlit==0.70.0
spacy-streamlit==1.0.0rc0
st-annotated-text==1.0.1
pandas==1.1.4
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.0.0a0/de_core_news_sm-3.0.0a0.tar.gz
